producer/requirements.txt
requests
python-dotenv
xmltodict
confluent-kafka